edited vector

+++ srcs/lib/vector/srcs/vector_pop_back.c
@@ -23,7 +23,7 @@ int		vector_pop_back(void **vector)
 	if (arr[2] - 1 < ft_ceill(arr[1] / A / A) && arr[1] != ARR_SIZE
 		&& arr[1] != ft_floorl(ARR_SIZE * A))
 	{
-		size = ft_ceill(arr[1] / A);
+		size = (size_t)ft_ceill(arr[1] / A);
 		if (!(tmp = (size_t *)malloc(sizeof(size_t) * 3 + size * arr[0])))
 			return (0);
 		ft_memcpy((void *)tmp, (void *)arr, arr[0] * arr[2] +
